Move actions to action folder
